API/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
tensorflow
psycopg2-binary
python-dotenv
selenium
